+++ packages/playhtml/vite.config.ts
@@ -11,8 +11,4 @@ export default defineConfig({
       fileName: (format) => `playhtml.${format}.js`,
     },
   },
-  // temp to deal with partykit double init on hot refresh
-  server: { hmr: false },
-  // Ignores `public` folder for the npm packge
-  publicDir: false,
 });